#393 move the elector and scheduler to new package
